Initialize the volume seekbar preference max and progress.
When the media stream is set, initialize the preference max and progress
with the streams current value. Otherwise, when we initialize the seekbar
volumizer, it will first set the seekbar max to 0 and progress to 0,
then update with the correct value, which will result in the jank that
is seen when the sound settings are displayed.

Change-Id: I515c97bbc6ec38bbe92755e3d7cb53bb13ac52d0
Fix: 34035654
Test: make RunSettingsRoboTests
(cherry picked from commit 9702262e0f44960df6af0a032c8a15c1534ecaae [formerly b7490bea28cb28d301fe4ec0275cef7129621206])

Former-commit-id: cbc231c5452e7eddfd26561746bd5c44dbdfb4ba