Reflect device encryption status
On the Settings & Location settings page, the
encryption_and_credential preference will statically show
"Phone encrypted" regardless of encryption status of phone.

This change changes Settings to check for encryption status
of the phone and update the summary as required.
This requires adding a string saying that the device is not encrypted.
And translations will have to happen after this is merged.

Change-Id: I2f5bd281d8f1ed32801446fcdd453ebbaa04bd32
Bug: 65535239
Fixes: 65535239
Test: manual, I had a gobo phone that was not encrypted,
made sure that it said "Phone not encrypted", then tap through
to encrypt phone, and then check that it is now "Phone encrypted".

Former-commit-id: f0c0320c82ab6c243596ede6b450dff5d1486df3