Explicitly set ENABLED state when a user enables an app.
When a user explicitly clicks a button labeled "Enable" for an
application in settings, the underlying package state should reflect
this. Otherwise, it is impossible to distinguish between the default
state of an app and an app which the user has explicitly requested to
enable.

This behavior is consistent with the Play Store's.

This also fixes a metrics bug where enabling a DISABLED_UNTIL_USED app
would be erroneously treated as disabling the app.

Test: Flashed and verified enabling both DISABLED_UNTIL_USED and
DISABLED apps.
Fixes: 64223206

Change-Id: Ic18068e6047045d6f5333eaf10fcddd1e95a679d

Former-commit-id: 3638973a1703b85a2fbed07bc5f221436126540d