Wifi: Restore hidden flag for hidden networks on update network
Currently when user modifies any of the network parameter for hidden
networks, hidden flag is not updated. As a result this network is no
longer part of hidden networks, we are unable to reconnect when user
triggers wifi off and on.

This commit sets the value of hiddenSSID flag of config at the Settings
side on network update.

Test: manual test with hidden access point
Bug: 73256783

Change-Id: Ic20f03345645d8bf3bc9c3ff44910dd3f2797fec

Former-commit-id: f44da638bdfc0c237627aae9e7affe0af723e375