Fix the crash when editing or deleting APN.
Previously ApnPrefence starts an intent with URI content://telephony/carrier/filtered/id,
this URI is only implemented in telephony provider for query(), but not
for delete() or update(). This caused the crash when user tries to
update or delete an APN through this URI.

We should let ApnPrefence starts an intent with URI content://telephony/carrier/id, which
is a general telephony URI and all of query() add() delete() update() are implemented for
non-DPC mode. And let ApnPrefence starts an intent with URI content://telephony/carrier/filtered/id
in DPC mode(as the general URI can't access DPC-owned APN), when only DPC-owned APNs are
presented to user. In the DPC mode, user can't update, add or delete an APN, they can
only view(query) the APN, so it won't crash with URI_FILTERED.

Bug: 72387301
Test: manual. Tried update, delete, and add action in Access Point Name page, no crash occurs. Instrumentation test will be added in b/72154761.
Change-Id: I9979cbffcc94a37b2bd96db766ececd0ac7b20e2

Former-commit-id: db9a9bb0b0438816909e002156cb7322c66a0d28